refactor: Count, Money vo 패키지로 이동, 코드 정렬
